Module quandl_v3::prelude

This crate's public interface.

This exclude error handling names to avoid needless conflicts with other crates (e.g. std::result::Result).

Structs

BatchQuery
Builder pattern run multiple queries in batch.

BatchQueryIterator
Iterator returned by the BatchQuery::run method.

Code
Data structure to hold the result of a code list query.

CodeListQuery
Query a list of dataset codes from a specific database.

DataAndMetadataQuery
Query the data and metadata from a specific dataset.

DataQuery
Query the data from a specific dataset.

DatabaseList
Data structure to hold the result of doing a search database query.

DatabaseMetadata
Hold the metadata associated to a specific database.

DatabaseMetadataQuery
Database metadata query.

DatabaseSearch
Query to search into a database metadata list.

DatasetList
Data structure to hold the result of a search dataset query.

DatasetMetadata
Hold the metadata associated to a specific dataset.

DatasetMetadataQuery
Dataset metadata query.

DatasetSearch
Query to search into a dataset metadata list.

SearchMetadata
Some queries, namely those which list datasets or databases metadata, often return some metadata about the search itself. This is a structure to hold that metadata.

Enums

Frequency
Parameters to indicate the desired frequency. When you change the frequency of a dataset, Quandl returns the last observation for the given period.

Order
Select the sort order with this enum. The default sort order is descending.

Transform
Perform calculations on your data prior to downloading.

Statics

QUANDL_API_URL
Quandl API URL used as the base URL for all queries.

Traits

ApiCall
Trait allowing implementers to submit a request through the Quandl API.

ApiParameters
Api parameters implemented by all queries.

DataParameters
Data parameters implemented by data fetching queries.

SearchParameters
Search parameters implemented by search queries.