Function quandl_v3::prelude::batch_query

pub fn batch_query<T, B, C>(queries: B, threads: usize) -> Iterator<Result<T>> where T: Decodable + Clone + Send + 'static, C: ApiCall<T> + Clone + Send + 'static, B: AsRef<[C]>

Submit multiple queries at the same time.

A slice/vector of queries is given as argument together with the number of threads that should be used. The number of threads will be truncated to the range [1, queries.as_ref().len()].

This function download the data from the Quandl servers asynchronously. It does so by returning an Iterator which return the Result from each individual query in the order they appear in the queries argument.