Struct quandl_v3::prelude::Data

pub struct Data<T: Decodable> {
    pub data: Vec<T>,
    pub column_names: Vec<String>,
    pub start_date: String,
    pub end_date: String,
    pub frequency: Frequency,
    pub order: Order,
    pub rows: Option<usize>,
    pub limit: Option<usize>,
    pub collapse: Option<Frequency>,
    pub transform: Option<Transform>,
    pub column_index: Option<usize>,
}

Data structure to hold the data resulting from a data query.

The data itself will be a tuple (or a Vector if all the elements are of the same type), but it is necessary to provide the type parameter explicitely for the underlying parser to know how to operate on the specific data received.

Quandl API Reference

Fields

data

The data requested.

column_names

A Vector containing the name of every column of the data.

start_date

The start date of the data as a String of the form format!("{:#4}-{:#2}-{:#2}", year, month, day);.

end_date

The end date of the data as a String of the form format!("{:#4}-{:#2}-{:#2}", year, month, day);.

frequency

The interval between two consecutive data point in the original dataset.

order

The ascending or descending ordering of the data (with respect to the date for time series).

rows

The number of rows to return. Is identical to the limit field.

limit

The number of rows to return. Is identical to the rows field.

collapse

The interval between two consecutive data point in the transformed (returned) dataset.

transform

The transform performed on this dataset; see also the Transform enum.

column_index

The column index that was returned. Note that the 'date' column (index 0) is always returned.