Struct quandl_v3::prelude::SearchMetadata

pub struct SearchMetadata {
    pub query: String,
    pub per_page: usize,
    pub current_page: usize,
    pub prev_page: Option<usize>,
    pub total_pages: usize,
    pub total_count: usize,
    pub next_page: Option<usize>,
    pub current_first_item: Option<usize>,
    pub current_last_item: Option<usize>,
}

Some queries, namely those which list datasets or databases metadata, often return some metadata about the search itself. This is a structure to hold that metadata.

Fields
query: String

A string of the search keywords submitted formatted as format!("{}+{}+...+{}", keyword_1, keyword_2, ..., keyword_n).

per_page: usize

The number of search result per page.

current_page: usize

The current page of result that was returned by this query.

prev_page: Option<usize>

The number of the previous page, unless there is no previous page.

total_pages: usize

The total number of pages that can be queried.

total_count: usize

The total number of search result returned.

next_page: Option<usize>

The number of the next page, unless there is no next page.

current_first_item: Option<usize>

Index of the first result on the current page, with respect to the total number of results.

current_last_item: Option<usize>

Index of the last result on the current page, with respect to the total number of results.